requirements.txt
pybit==5.6.2
python-dotenv==1.0.0
requests==2.31.0